Added a new object to handle the game state instead
